feat: Integrated and refactored image recognition part of the app
