--- Resources/Paths.py
@@ -1,7 +1,0 @@
-from pathlib import Path
-
-haarcascade = str(Path(__file__).parents[1] / "Resources/haarcascade_frontalface_default.xml")
-test_photo = str(Path(__file__).parents[1] / "Assets/test.jpg")
-shrek_ear = str(Path(__file__).parents[1] / "Assets/shrek_ear.png")
-output_path = str(Path(__file__).parents[1] / "Assets/output.png")
-
