+++ app/assets/shrek_ear.png
