+++ app/models.py
@@ -1,5 +1,6 @@
 from sqlalchemy import Column, Integer, String
 from app.database import base
+from pydantic import BaseModel
 
 
 class Images(base):
@@ -7,3 +8,7 @@ class Images(base):
 
     id = Column(Integer, primary_key=True, index=True)
     url = Column(String)
+
+
+class Image(BaseModel):
+    url: str
