+++ app/resources/constants.py
@@ -1,5 +1,3 @@
-
-
 right_width_multiplier = 2 / 3
 left_width_multiplier = 3 / 12
-heigth_divider = 2
+height_divider = 2
