+++ requirements.txt
@@ -17,6 +17,5 @@ starlette==0.20.4
 toml==0.10.2
 typing_extensions==4.4.0
 uvicorn==0.19.0
-
 Pillow~=9.3.0
 opencv-python~=4.6.0.66
